+++ src/navigations/ExchangeStack.tsx
@@ -1,12 +1,12 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import GPS from "../screens/ExchangeView/GPS";
+import GPSView from "../screens/Exchange/GPSView";
 
 const Stack = createNativeStackNavigator();
 
 export default function ExchangeStack() {
     return (
         <Stack.Navigator>
-            <Stack.Screen name="명함 교환" component={GPS} />
+            <Stack.Screen name="명함 교환" component={GPSView} />
         </Stack.Navigator>
     )
 };
