+++ src/screens/CardList/CardDetailView.tsx
