+++ src/screens/CardList/CardListView.tsx
@@ -1,0 +1,9 @@
+import { Text, View } from 'react-native';
+
+export default function CardListView({ navigation }: any) {
+	return (
+		<View>
+			<Text>명함 리스트</Text>
+		</View>
+	);
+}
