--- src/screens/CardListView/CardList.tsx
@@ -1,9 +1,0 @@
-import { Text, View, Button } from 'react-native';
-
-export default function CardList({ navigation }: any) {
-	return (
-		<View>
-			<Text>명함 리스트</Text>
-		</View>
-	);
-}
