+++ src/screens/Exchange/GPSView.tsx
@@ -1,6 +1,6 @@
 import { Text } from "react-native";
 
-export default function GPS() {
+export default function GPSView() {
     return (
         <Text>
             웹소켓 교환
