+++ src/screens/MyPage/CardCreateView.tsx
@@ -1,6 +1,6 @@
 import { Text } from "react-native";
 
-export default function CardCreate() {
+export default function CardCreateView() {
     return (
         <Text>
             명함 생성
